+++ webpack.config.base.js
@@ -14,6 +14,7 @@ module.exports = {
             {test: /\.eot(\?v=\d+\.\d+\.\d+)?$/, loader: 'file'},
             {test: /\.svg(\?v=\d+\.\d+\.\d+)?$/, loader: 'url?limit=10000&mimetype=image/svg+xml'},
             {test: /\.css$/, loader: 'style!css'},
+            {test: /\.less/, loader: 'style!css!less'},
             {test: /\.html/, loader: 'raw'}
         ]
     },